structure: move client-needed types to shared/
Art type constants move from server/art/types.ts to shared/art/types.ts,
re-exported by server/art so server consumers are unchanged. DailyPlaytime
moves to shared/types/Activity.ts. The Steam art debug page swaps the full
SteamPicsMetadata row type for a narrow shared/types/SteamArt.ts type
covering only the fetchedAt and path/hash columns it renders.
