+++ server/art/invalidate.ts
@@ -1,6 +1,6 @@
 import fs from "node:fs";
+import type { ArtKey } from "#shared/art/types";
 import { artDirectory } from "./paths";
-import type { ArtKey } from "./types";
 
 // Drops every cached file and `.missing` marker for a game, so the next render
 // re-resolves its sources. Used when upstream art paths change.
