+++ shared/types/Activity.ts
@@ -1,0 +1,4 @@
+export interface DailyPlaytime {
+  date: string;
+  minutes: number;
+}
